Update JAX to 0.4.29
This was a big headache. Updating JAX entails updating cuDNN, and the version
of cuDNN that poetry initially picked has a bug that caused incorrect samples
(maybe also incorrect other stuff) with batch size 64 but not with batch size 1
or 2. Pinned a specific, newer cuDNN and wrote a test for this.